pyproject.toml
[tool.poetry]
name = "xray-node"
version = "0.0.1"
description = ""
authors = ["laoshan-taoist <65347330+laoshan-taoist@users.noreply.github.com>"]
license = "AGPLv3"
packages = [
    { include = "xray_node" },
]
readme = "README.md"
repository = "https://github.com/laoshan-tech/xray-node"
homepage = "https://github.com/laoshan-tech/xray-node"

[tool.poetry.dependencies]
python = "^3.8"
httpx = "^0.17.1"
psutil = "^5.8.0"
tomlkit = "^0.7.0"
grpcio = "^1.37.0"
grpcio-tools = "^1.37.0"
xray-rpc = "^1.4.2"
uvloop = [
    { version = "^0.15.2", platform = "linux|macos" },
]
grpcio-status = "^1.37.0"
click = "^7.1.2"
tortoise-orm = "^0.17.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest-asyncio = "^0.14.0"
pytest-ordering = "^0.6"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.poetry.scripts]
xnode = "xray_node.main:main"

[tool.black]
line-length = 120
target-version = ["py38"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
